Move actor context obtaining to `Events`
